requirements.txt
Flask
vaderSentiment
spotipy
openai
python-dotenv